changes made to the app

+++ app/controllers/interest_controller.rb
@@ -16,24 +16,32 @@ class InterestController < ApplicationController
   
   def add_interest
     @user = current_user
-    
+    @league_interest = []
+    unless @user.leagues.blank?
+      for league in @user.leagues    
+        @league_interest << league.interest 
+      end
+    end
     if request.post?
-       
-       for interest in @user.user_interests 
-         interest.destroy
+     unless @user.user_interests.blank?     
+       for interest in @user.user_interests
+         unless @league_interest.blank?
+           unless @league_interest.include?(interest.interest)
+            interest.destroy
+           end
+         end
        end
-       
-       unless params[:interest_ids].blank?
-            UserInterest.create(interest_id:1, user_id:current_user.id)
+      end
+      @interest = Interest.where(:interest_name=>"Random").last
+      unless params[:interest_ids].blank?
+           UserInterest.create(interest_id:@interest.id, user_id:current_user.id)
            for interest_id in params[:interest_ids]
               UserInterest.create(interest_id: interest_id, user_id:current_user.id)
            end
        end
        flash[:notice] = "Interest list has been updated"
        redirect_to stage_path
-    end
-    
-    
+    end    
   end
     
   def resolve_layout
